Reduce publication test refactor churn
